Move cloud_storage.rb up to scripts/
