+++ scripts/brew/artifact-meta/dump.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-require_relative "../cloud_storage.rb"
+require_relative "../../cloud_storage.rb"
 require_relative "../tap.rb"
 require_relative "../cask.rb"
 require "fileutils"
