+++ scripts/scoop/update-times/dump.rb
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 
-require_relative "../cloud_storage.rb"
+require_relative "../../cloud_storage.rb"
 require_relative "../bucket.rb"
 require "fileutils"
 require "json"
